+++ app/src/main/java/com/example/artworksmanager/ui/collection/CollectionFragment.kt
@@ -30,7 +30,7 @@ class CollectionFragment : Fragment() {
 
     private val adapter = ArtworkAdapter { artwork ->
         findNavController().navigate(
-            CollectionFragmentDirections.actionCollectionToDetail(artwork.id)
+            CollectionFragmentDirections.actionCollectionToDetail(artwork.id.toInt())
         )
     }
